Rajout des derniers tests et des contraintes graphiques

+++ P10_01_Xcode/Model/Network/RecipeDecoder.swift
@@ -42,7 +42,7 @@ struct RecipeDecoder: Decodable {
         let url: URL
         let ingredients: [Ingredient]
         let totalTime: Double
-        let yield: Int
+        let yield: Double
     }
     
     //
